integration tests: move complex scenario-based integration tests to a separate file
